refactor(backend): finalized arch and decided on verbose sync + async approach
Since most of the code for fastapi is async, I was trying to keep my approaches to integrating the celery worker code DRY and keep similar resource managers like database, redis, and http only async. In a seperate sandbox enviroment, I was testing around with celery in async contexts, trying to do complex async event loop management and using shared resources, however ultimately failed trying to cram something into celery that is not inheritely built in. In order to stick to celery's strengths, I decided on biting the bullet and making the codebase more verbose with Async / Sync Classes for managers and for somerepositories. in the end though, I believe this will be what is best for the codebase, and although its a bit more tedious for testing and writing out, the clear seperations should assist when the two contexts are required. only implementing what I need though for both i.e. in the sync user repo, only implementing functions needed during scripts / celery tasks in sync context. vice versa for async, only functions needed for the fastapi.